requirements.txt
Flask
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
spotipy
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
